fixed two warnings

+++ src/fonts/vertex-buffer.c
@@ -229,8 +229,8 @@ vertex_buffer_print( vertex_buffer_t * self )
 
     assert(self);
 
-    fprintf( stderr, "%ld vertices, %ld indices\n",
-             vector_size( self->vertices ), vector_size( self->indices ) );
+    fprintf( stderr, "%d vertices, %d indices\n",
+             (int)vector_size( self->vertices ), (int)vector_size( self->indices ) );
     while( self->attributes[i] )
     {
         int j = 8;
